BACKLOG-19921: reorganized remaining configs
