+++ src/javascript/SelectorTypes/Picker/PickerDialog/RightPanel/PickerSelection/Selection.scss
@@ -9,6 +9,10 @@
     color: var(--color-gray60);
 }
 
+.gray {
+    color: var(--color-gray);
+}
+
 .text {
     gap: var(--spacing-small);
     width: 100%;
