--- src/javascript/SelectorTypes/Picker/accordionItems/ContentTypeSelector/index.js
@@ -1,4 +1,0 @@
-import {EditorialLinkContentTypeSelector} from './ContentTypeSelector';
-import {ContentTypeSelector as JContentTypeSelector} from '@jahia/jcontent';
-
-export {EditorialLinkContentTypeSelector, JContentTypeSelector};
