+++ src/javascript/SelectorTypes/Picker/configs/categoryPicker.js
@@ -2,8 +2,8 @@ import {Constants} from '~/SelectorTypes/Picker/Picker2.constants';
 import {mergeDeep} from '~/SelectorTypes/Picker/Picker2.utils';
 import {ContentPickerConfig} from '~/SelectorTypes/Picker/configs/ContentPickerConfig';
 import {Collections} from '@jahia/moonstone';
-import {PickerTreeQueryHandler} from '~/SelectorTypes/Picker/accordionItems/QueryHandlers/queryHandlers';
-import {renderer} from '~/SelectorTypes/Picker/accordionItems/renderer';
+import {PickerTreeQueryHandler} from '~/SelectorTypes/Picker/configs/queryHandlers';
+import {renderer} from '~/SelectorTypes/Picker/configs/renderer';
 import React from 'react';
 
 export const registerCategoryPicker = registry => {
