+++ src/javascript/SelectorTypes/Picker/configs/editorialLinkPicker/EditorialLinkContentTypeSelector.scss
