+++ src/javascript/SelectorTypes/Picker/configs/editorialLinkPicker/index.js
@@ -1,0 +1,1 @@
+export * from './editorialLinkPicker';
