+++ src/javascript/SelectorTypes/Picker/configs/editorialPicker.js
@@ -1,0 +1,79 @@
+import {Constants} from '~/SelectorTypes/Picker/Picker2.constants';
+import {mergeDeep} from '~/SelectorTypes/Picker/Picker2.utils';
+import {ContentPickerConfig} from '~/SelectorTypes/Picker/configs/ContentPickerConfig';
+import {getPagesSearchContextData} from '~/SelectorTypes/Picker/configs/getPagesSearchContextData';
+import {PickerContentsFolderQueryHandler, PickerPagesQueryHandler} from '~/SelectorTypes/Picker/configs/queryHandlers';
+import {renderer} from '~/SelectorTypes/Picker/configs/renderer';
+import React from 'react';
+import {registry} from '@jahia/ui-extender';
+import {ContentTypeSelector as JContentTypeSelector, ViewModeSelector} from '@jahia/jcontent';
+import {
+    cePickerSetPage,
+    cePickerSetTableViewMode,
+    cePickerSetTableViewType
+} from '~/SelectorTypes/Picker/Picker2.redux';
+
+const viewModeSelectorProps = {
+    selector: state => ({
+        mode: state.contenteditor.picker.mode,
+        viewMode: state.contenteditor.picker.tableView.viewMode
+    }),
+    setTableViewModeAction: mode => cePickerSetTableViewMode(mode)
+};
+
+const contentTypeSelectorProps = {
+    selector: state => ({
+        mode: state.contenteditor.picker.mode,
+        siteKey: state.site,
+        path: state.contenteditor.picker.path,
+        lang: state.language,
+        uilang: state.uilang,
+        params: {
+            selectableTypesTable: registry.get('pickerConfiguration', state.contenteditor.picker.pickerKey)?.selectableTypesTable || []
+        },
+        pagination: state.contenteditor.picker.pagination,
+        sort: state.contenteditor.picker.sort,
+        tableView: state.contenteditor.picker.tableView
+    }),
+    reduxActions: {
+        setPageAction: page => cePickerSetPage(page),
+        setTableViewTypeAction: view => cePickerSetTableViewType(view)
+    }
+};
+
+export const registerEditorialPicker = registry => {
+    registry.add(Constants.pickerConfig, 'editorial', mergeDeep({}, ContentPickerConfig, {
+        searchSelectorType: 'jmix:searchable',
+        selectableTypesTable: ['jnt:page', 'jnt:contentList', 'jnt:contentFolder', 'jmix:siteContent', 'jmix:editorialContent']
+    }));
+
+    // These are jcontent accordion items, additional targets are added to enhance selection
+    const pagesItem = registry.get(Constants.ACCORDION_ITEM_NAME, Constants.ACCORDION_ITEM_TYPES.PAGES);
+    const contentFoldersItem = registry.get(Constants.ACCORDION_ITEM_NAME, Constants.ACCORDION_ITEM_TYPES.CONTENT_FOLDERS);
+
+    if (pagesItem) {
+        // Page content
+        registry.add(Constants.ACCORDION_ITEM_NAME, `picker-${Constants.ACCORDION_ITEM_TYPES.PAGES}`, {
+            ...pagesItem,
+            viewSelector: <ViewModeSelector {...viewModeSelectorProps}/>,
+            tableHeader: <JContentTypeSelector {...contentTypeSelectorProps}/>,
+            targets: ['default:50', 'editorial:50'],
+            defaultSort: {orderBy: 'lastModified.value', order: 'DESC'},
+            getSearchContextData: getPagesSearchContextData,
+            queryHandler: PickerPagesQueryHandler
+        }, renderer);
+    } else {
+        console.warn('Picker will not function properly due to missing accordionItem for pages');
+    }
+
+    if (contentFoldersItem) {
+        registry.add(Constants.ACCORDION_ITEM_NAME, `picker-${Constants.ACCORDION_ITEM_TYPES.CONTENT_FOLDERS}`, {
+            ...contentFoldersItem,
+            viewSelector: <ViewModeSelector {...viewModeSelectorProps}/>,
+            targets: ['default:60', 'editorial:60'],
+            queryHandler: PickerContentsFolderQueryHandler
+        }, renderer);
+    } else {
+        console.warn('Picker will not function properly due to missing accordionItem for content-folders');
+    }
+};
