+++ src/javascript/SelectorTypes/Picker/configs/mediaPicker/MediaPickerConfig.js
