+++ src/javascript/SelectorTypes/Picker/configs/mediaPicker/MediaPickerConfig.spec.js
