+++ src/javascript/SelectorTypes/Picker/configs/mediaPicker/index.js
@@ -1,0 +1,2 @@
+export * from './mediaPicker';
+export {FilePickerCaption} from '~/SelectorTypes/Picker/configs/mediaPicker/FilePickerCaption';
