+++ src/javascript/SelectorTypes/Picker/configs/renderer.jsx
