+++ src/javascript/SelectorTypes/Picker/configs/usergroupPicker.js
@@ -1,9 +1,9 @@
 import {Constants} from '~/SelectorTypes/Picker/Picker2.constants';
 import {mergeDeep} from '~/SelectorTypes/Picker/Picker2.utils';
 import {ContentPickerConfig} from '~/SelectorTypes/Picker/configs/ContentPickerConfig';
-import {transformQueryHandler} from '~/SelectorTypes/Picker/accordionItems/QueryHandlers/queryHandlers';
+import {transformQueryHandler} from '~/SelectorTypes/Picker/configs/queryHandlers';
 import {Group} from '@jahia/moonstone';
-import {renderer} from '~/SelectorTypes/Picker/accordionItems/renderer';
+import {renderer} from '~/SelectorTypes/Picker/configs/renderer';
 import React from 'react';
 import {reactTable, Sql2SearchQueryHandler} from '@jahia/jcontent';
 import {UserGroupPickerFragment} from './usergroupPicker.gql-queries';
