+++ src/javascript/SelectorTypes/Picker/registerPicker.js
@@ -1,7 +1,6 @@
 import {registerPickerConfig} from './configs/Picker2.configs';
 import {registerPickerActions} from './actions/registerPickerActions';
 import {Picker2} from './Picker2';
-import {registerAccordionItems} from '~/SelectorTypes/Picker/accordionItems/accordionItems';
 import {registerPickerReducer} from '~/SelectorTypes/Picker/Picker2.redux';
 import {Constants} from '~/SelectorTypes/Picker/Picker2.constants';
 
@@ -33,6 +32,5 @@ export const registerPicker = registry => {
         resolver: options => getPickerSelectorType(registry, options)
     });
     registerPickerActions(registry);
-    registerAccordionItems(registry);
     registerPickerReducer(registry);
 };
